backend-api/requirements.txt
fastapi[all]
uvicorn[standard]
google-cloud-pubsub
python-dotenv
click